Refactored scope module. Implemented filter in *ListApiHandlers
